+++ DQQ/DQQ.Core/Profiles/Items/Equipments/TwoHandWeapons/CorrodedBlade.cs
@@ -21,7 +21,7 @@ namespace DQQ.Profiles.Items.Equipments.TwoHandWeapons
     public override string? Discription => "Corroded Blade";
 
     public override decimal AttackPerSecond => 1m;
-    public override BigInteger BaseDamage => 10;
+    public override Int64 BaseDamage => 10;
 
     public override EquipComponent GenerateEquipComponent(int? itemLevel)
     {
