+++ DQQ/DQQ.Core/Profiles/Items/Equipments/WeopnProfile.cs
@@ -5,6 +5,6 @@ namespace DQQ.Profiles.Items.Equipments
   public abstract class WeopnProfile : EquipProfile
   {
     public abstract decimal AttackPerSecond { get; }
-    public abstract BigInteger BaseDamage { get; }
+    public abstract Int64 BaseDamage { get; }
   }
 }
